Re-add LocalVLLMServer and enable VCR
