+++ src/granite_io/backend/__init__.py
@@ -11,3 +11,4 @@ from granite_io.backend.registry import backend, make_backend  # noqa: F401
 import granite_io.backend.litellm  # noqa: F401
 import granite_io.backend.openai  # noqa: F401
 import granite_io.backend.transformers  # noqa: F401
+import granite_io.backend.vllm_server  # noqa: F401
