+++ tests/io/test_certainty.py
@@ -8,11 +8,16 @@ Tests for the Granite certainty intrinsic's I/O processor
 import datetime
 import textwrap
 
+# Third Party
+import pytest
+
 # Local
 from granite_io import make_io_processor
+from granite_io.backend.vllm_server import LocalVLLMServer
 from granite_io.io.certainty import CertaintyCompositeIOProcessor, CertaintyIOProcessor
 from granite_io.io.granite_3_2.input_processors.granite_3_2_input_processor import (
     Granite3Point2Inputs,
+    override_date_for_testing,
 )
 from granite_io.types import (
     GenerateResult,
@@ -108,7 +113,8 @@ def test_canned_output():
     assert multi_output_strs == multi_expected
 
 
-def test_run_model(lora_server):
+@pytest.mark.vcr
+def test_run_model(lora_server: LocalVLLMServer, fake_date: str):
     """
     Run a chat completion through the LoRA adapter using the I/O processor.
     """
@@ -116,13 +122,15 @@ def test_run_model(lora_server):
     io_proc = CertaintyIOProcessor(backend)
 
     # Pass our example input thorugh the I/O processor and retrieve the result
+    override_date_for_testing(fake_date)  # For consistent VCR output
     chat_result = io_proc.create_chat_completion(_EXAMPLE_CHAT_INPUT)
 
     # We run at temperature zero, so this result should be consistent
     assert float(chat_result.results[0].next_message.content) == 0.8
 
 
-def test_run_composite(lora_server):
+@pytest.mark.vcr
+def test_run_composite(lora_server: LocalVLLMServer, fake_date: str):
     """
     Generate chat completions and check certainty using a composite I/O processor to
     choreograph the flow.
@@ -138,6 +146,7 @@ def test_run_composite(lora_server):
     input_without_msg = _EXAMPLE_CHAT_INPUT.model_copy(
         update={"messages": _EXAMPLE_CHAT_INPUT.messages[:-1]}
     ).with_addl_generate_params({"temperature": 0.2, "n": 5})
+    override_date_for_testing(fake_date)  # For consistent VCR output
     results = io_proc.create_chat_completion(input_without_msg)
     assert len(results.results) > 1
